backend_django/requirements.txt
django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
Pillow>=10.0
flask>=3.0
flask-cors>=4.0
requests>=2.31.0
